LayoutTests/http/tests/media/media-stream/disconnected-frame-permission-denied-expected.txt

frame "<!--framePath //<!--frame0-->-->" - has 1 onunload handler(s)
Tests that when a getUserMedia request is made, permission is denied and its Frame is disconnected before a callback is made, the error callback is invoked with PERMISSION_DENIED.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS error.name is "NotAllowedError"

PASS No callbacks invoked
PASS successfullyParsed is true

TEST COMPLETE
